requirements.txt
asgiref==3.5.2
Django==4.0.6
Pillow==9.2.0
PyMySQL==1.0.2
sqlparse==0.4.2
urllib3==1.26.9
astroid==2.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
django-ckeditor==6.3.2
django-js-asset==2.0.0
dynaconf==3.1.7
idna==3.3
isort==5.10.1
jsonfield==3.1.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
oauthlib==3.2.0
platformdirs==2.5.1
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
toml==0.10.2
typing-extensions==4.1.1
wrapt==1.13.3